+++ src/Badgified.cpp
@@ -1,18 +1,21 @@
-#include <alphalaneous.badgify/include/Badgify.hpp>
+#include <Geode/Geode.hpp>
+#include <dasshu.badgified/include/Badgified.hpp>
+
 #include "ModHelper.hpp"
 
 using namespace geode::prelude;
+using namespace dasshu::badgified;
 
 $on_mod(Loaded) {
-    if (!ModHelper::isBadgifyLoaded())
+    if (ModHelper::isBadgifiedLoaded() && ModHelper::isBadgifyLoaded())
         return;
 
     for (auto const& [id, info] : ModHelper::badges) {
-        alpha::badgify::registerBadge(
+        registerBadge(
             fmt::format("{}"_spr, id),
             info.name,
             info.description,
-            [id, info](alpha::badgify::Badge const& badge) {
+            [id, info](Badge const& badge) {
                 if (id == "star-collector" && badge.user.data()->m_stars < 3000) return;
                 if (id == "moon-collector" && badge.user.data()->m_moons < 3000) return;
                 if (id == "secret-coin-collector" && badge.user.data()->m_secretCoins < 100) return;
@@ -25,7 +28,7 @@ $on_mod(Loaded) {
                 
                 if (id == "creator" && badge.user.data()->m_creatorPoints == 0) return;
 
-                alpha::badgify::showBadge(
+                showBadge(
                     badge,
                     CCSprite::createWithSpriteFrameName(info.sprite)
                 );